% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{Clusterec}
\alias{Clusterec}
\title{Cluster equivalence class}
\usage{
Clusterec(ec)
}
\arguments{
\item{ec}{A \code{list} of equivalence class}
}
\value{
A \code{list}. Each element is an ec cluster.
}
\description{
Find equivalence class clusters
}
\details{
Cluster ec with common transcripts.
}
\examples{
require('magrittr')

ec <- system.file('extdata', 'example.ec', package = 'RNASeqEM') \%>\% read_ec
Clusterec(ec)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
