% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{collapseL_}
\alias{collapseL_}
\title{Internal functions for cluster EM}
\usage{
collapseL_(l, logidx)
}
\arguments{
\item{l}{A \code{list}. Each element is a vector.}
}
\value{
A \code{list}.
}
\description{
Cluster EM internal functions.
}
\details{
Collapse list.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
