% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM}
\alias{EM}
\title{EM model}
\usage{
EM(efflenraw, ecraw, countraw, spenum, maxiter = 10000L, miniter = 50L,
  details = FALSE)
}
\arguments{
\item{efflenraw}{A code{arma::vec} indicating the effect length of transcript.}

\item{ecraw}{A \code{character vector} and each element is a string with comma delimiter.}

\item{countraw}{A \code{arma::uvec} indicates the counts of ec.}

\item{spenum}{A \code{arma::uvec} indicated the transcript number in each species.}

\item{maxiter}{The maximum iteration number with the default value of 10000.}

\item{miniter}{The minimum iteration number with the default value of 50.}

\item{details}{A \code{bool} value.  When it is set as \code{true}, logistic likelihood and counts for each species in every iteration will be returned, otherwise \code{false}.}
}
\value{
A \code{List} indicates estimated counts of transcripts.
}
\description{
Expectation maximization (EM) model for RNA-seq quantification.
}
\details{
EM model for RNA-seq quantification. The equivalence class (ec) with 0 counts are removed, because these counts have no contributes to the final results.
}
\examples{
## Single species
##    f1 f2 f3
## ec1 1 1 1
## ec2 0 1 1
## ec3 1 0 1
## ec4 1 0 0
## ec5 1 1 0
plist <- list(ec = c('0,1,2', '1,2', '0,2', '0', '0,1'), count = rep(1, 5), efflen = rep(1, 3))
EM(plist$efflen, plist$ec, plist$count, spenum = 3)

## Two species
##    f1 f2 f3 f1' f2'
## ec1 1  1  0  0  1
## ec2 1  0  1  1  0
## ec3 0  1  1  0  0
## ec4 0  0  0  1  1
## ec5 1  0  1  0  1
## ec6 1  1  0  0  0
plist <- list(ec = c('0,1,4', '0,2,3', '1,2', '3,4', '0,2,4', '0,1'),
              count = rep(1, 6), efflen = rep(1, 5))
EM(plist$efflen, plist$ec, plist$count, c(3, 2))
## compare with single species
EM(plist$efflen, plist$ec, plist$count, 5)
}
\references{
\href{https://arxiv.org/abs/1104.3889}{Lior Pachter: Models for transcript quantification from RNA-Seq}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
