% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMSingle}
\alias{EMSingle}
\title{Single EM iteration}
\usage{
EMSingle(countper, efflen, ec, ecnum, spenum)
}
\arguments{
\item{countper}{A \code{arma::vec} of input count percentage.}

\item{efflen}{A \code{std::vector<arma::vec>} indicated the effective length of transcripts.}

\item{ec}{A \code{std::vector<arma::vec>} indicated equivalence classes with the same length of \code{efflen}.}

\item{ecnum}{A \code{arma::uvec} indicated the mapped count number of equivalence class with the same length of \code{efflen}.}

\item{spenum}{A \code{arma::uvec} indicated number of transcripts.}
}
\value{
A updated \code{arma::vec} of count percentage.
}
\description{
Parallel a single EM iteration.
}
\details{
Add zero at the head of input \code{spenum}.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
