% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estcount2Prob}
\alias{Estcount2Prob}
\title{Counts to probabilities}
\usage{
Estcount2Prob(estcount, spenum)
}
\arguments{
\item{estcount}{A \code{arma::vec} estimated counts of transcripts.}

\item{spenum}{A \code{arma::uvec} indicated number of transcripts.}
}
\value{
A \code{arma::vec} indicates probabilities of selecting a read from the different transcripts.
}
\description{
Transform estimated count to probabilities.
}
\details{
Use estimated counts as the outputs and EM stop conditions.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
