% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientSM}
\alias{GradientSM}
\alias{GradientSP}
\alias{GradientISRU}
\title{Calculate gradient}
\usage{
GradientSM(w, efflen, ec, count, idx)

GradientSP(w, efflen, ec, count, idx)

GradientISRU(w, efflen, ec, count, alpha, idx)
}
\arguments{
\item{w}{A \code{arma::vec} indicates estimated parameters.}

\item{efflen}{A \code{std::vector<arma::vec>} indicated effective length of transcripts.}

\item{ec}{A \code{std::vector<arma::uvec>} indicated equivalence classes (ec).}

\item{count}{A \code{arma::uvec} indicated counts of ec.}

\item{idx}{A \code{arma::uvec} indicates the index of \code{w} used for gradient descending.}

\item{alpha}{\code{double}.}
}
\value{
A \code{arma::vec} indicates gradients.
}
\description{
Gradient.
}
\details{
\itemize{
  \item \code{GradientSM()}: Gradients of Softmax.
  \item \code{GradientSP()}: Gradients of SoftPlus.
  \item \code{GradientISRU()}: Gradients of ISRU.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
