% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IdxSpenum}
\alias{IdxSpenum}
\title{Index species number}
\usage{
IdxSpenum(spenum)
}
\arguments{
\item{spenum}{A \code{arma::uvec} indicated the transcript number in each species.}
}
\value{
A \code{arma::uvec}.
}
\description{
Index transcripts number of input species.
}
\details{
Add zero at the head of input \code{spenum}.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
