% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isEqualStr}
\alias{isEqualStr}
\title{Two strings comparison}
\usage{
isEqualStr(str1, str2)
}
\arguments{
\item{str1}{str2 \code{std::string} strings.}
}
\value{
A \code{bool} indicating whether two strings are equal.
}
\description{
Compare two strings
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
