% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InvSqrtRoot}
\alias{InvSqrtRoot}
\alias{ISRU1}
\alias{ISRU}
\alias{ISRUGrad1}
\alias{ISRUGrad}
\title{ISRU}
\usage{
InvSqrtRoot(x, alpha)

ISRU1(x, isr, alpha)

ISRU(x, isr, weight, alpha)

ISRUGrad1(x, isr, alpha)

ISRUGrad(x, isr, weight, alpha)
}
\arguments{
\item{x}{A \code{arma::vec}.}

\item{alpha}{\code{double}.}

\item{isr}{\code{arma::vec} indicating the inverse square root unit.}

\item{weight}{A \code{arma::vec} indicating the weight.}
}
\value{
\itemize{
  \item \code{InvSqrtRoot()}: A \code{arma::vec} indicates the inverse square root \eqn{\frac{1}{sqrt(1+\alpha x^2)}}.
  \item \code{ISRU()} and \code{ISRU1()}: A \code{arma::vec} indicates ISRU with (\eqn{(\frac{x_i}{sqrt(1+\alpha x_i^2)} + \frac{1}{\sqrt(\alpha)}) * weight_i} or without weight (\eqn{\frac{x_i}{sqrt(1+\alpha x_i^2)} + \frac{1}{\sqrt(\alpha)}}).
  \item \code{ISRUGrad()} and \code{ISRUGrad1()}: A \code{arma::vec} indicates \eqn{(\frac{1}{sqrt(1+\alpha x_i^2)})^3} or \eqn{(\frac{1}{sqrt(1+\alpha x_i^2)})^3 * weight_i}.
}
}
\description{
Calculate the Inverse square root unit (ISRU)
}
\details{
\itemize{
  \item \code{InvSqrtRoot()}: Inverse square root.
  \item \code{ISRU()} and \code{ISRU1()}: ISRU with or without weight.
  \item \code{ISRUGrad()} and \code{ISRUGrad1()}: Internal functions of Partial derivation.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
