% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LLEM}
\alias{LLEM}
\title{Calculate log-likelihood}
\usage{
LLEM(prob, efflen, ec, count)
}
\arguments{
\item{prob}{A \code{arma::vec} indicates probabilities of selecting a read from the different transcripts.}

\item{efflen}{A \code{std::vector<arma::vec>} indicated effective length of transcripts.}

\item{ec}{A \code{std::vector<arma::uvec>} indicated equivalence classes (ec).}

\item{count}{A \code{arma::uvec} indicated counts of ec.}
}
\value{
A \code{double} indicates log-likelihood.
}
\description{
Logistic likelihood.
}
\details{
The log-likelihood of given probabilities.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
