% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Max}
\alias{Max}
\title{Extract maximum elements}
\usage{
Max(vec1, vec2)
}
\arguments{
\item{vec1}{vec2 \code{arma::vec} vectors.}
}
\value{
A \code{arma::vec} vector indicating maximum element by pair-wise comparison.
}
\description{
Pairwise maximum elements of two vectors
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
