% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{read_ec}
\alias{read_ec}
\title{Standard read in ec files}
\usage{
read_ec(ecpath)
}
\arguments{
\item{ecpath}{The path of ec files.}
}
\value{
A \code{list}. 1st element is ec, and 2nd element is transcript ID.
}
\description{
Read in equivalence class
}
\details{
Read in equivalence class file generated from \code{kallisto} (version 0.44.0). The \code{*.ec} file can be generated by the \code{pseudo} command in the \code{kallisto}.
}
\examples{
require('magrittr')

ec <- system.file('extdata', 'example.ec', package = 'RNASeqEM') \%>\% read_ec
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
