% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogSumExp}
\alias{LogSumExp}
\alias{LogSumExp1}
\alias{Softmax}
\alias{Softmax1}
\title{Softmax}
\usage{
LogSumExp(x, weight)

LogSumExp1(x)

Softmax(x, weight)

Softmax1(x)
}
\arguments{
\item{x}{A \code{arma::vec}.}

\item{weight}{A \code{arma::vec} indicating the weight.}
}
\value{
\itemize{
  \item \code{LogSumExp()} and \code{LogSumExp1()}: A \code{double} indicates log-sum-exp.
  \item \code{Softmax()}: A \code{arma::vec} indicates the \eqn{\frac{\mathrm{e}^{x_i} * weight_i}{\sum{\mathrm{e}^{x_i} * weight_i}}}.
  \item \code{Softmax1()}: A \code{arma::vec} indicates the \eqn{\frac{\mathrm{e}^{x_i}}{\sum{\mathrm{e}^{x_i}}}}.
}
}
\description{
Calculate the log-sum-exp and softmax calculator
}
\details{
\itemize{
  \item \code{LogSumExp()} and \code{LogSumExp1()}: log-sum-exp with or without weight.
  \item \code{Softmax()} and \code{Softmax1()}: Softmax with or without weight.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
