% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Logistic}
\alias{Logistic}
\alias{Softplus1}
\alias{Softplus}
\alias{SoftplusGrad1}
\alias{SoftplusGrad}
\title{Softplus}
\usage{
Logistic(x)

Softplus1(x)

Softplus(x, weight)

SoftplusGrad1(x)

SoftplusGrad(x, weight)
}
\arguments{
\item{x}{A \code{arma::vec}.}

\item{weight}{A \code{arma::vec} indicating the weight.}
}
\value{
\itemize{
  \item \code{Logisitc()}: A \code{arma::vec} indicates the logistic function \eqn{\frac{1}{1 + \mathrm{e}^{-x}}}.
  \item \code{Softplus()} and \code{Softplus1()}: A \code{arma::vec} indicates softplus with (\eqn{\log(1 + \mathrm{e}^{x_i}) * weight_i} or without weight (\eqn{\log(1 + \mathrm{e}^{x_i})}).
  \item \code{SoftplusGrad()} and \code{SoftplusGrad1()}: A \code{arma::vec} indicates \eqn{\frac{logistic * weight_i}{\sum{softplus}}} or \eqn{\frac{logistic}{\sum{softplus1}}}.
}
}
\description{
Calculate the logistic and softplus calculator
}
\details{
\itemize{
  \item \code{Logistic()}: The logistic function.
  \item \item \code{Softplus()} and \code{Softplus1()}: Softplus with or without weight.
  \item \code{SoftplusGrad()} and \code{SoftplusGrad1()}: Internal functions for partial derivation.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
