% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpeCount}
\alias{SpeCount}
\title{Species estimated counts}
\usage{
SpeCount(est, spenum)
}
\arguments{
\item{est}{A \code{arma::vec} estimated counts of each transcripts.}

\item{spenum}{A \code{arma::uvec} indicated the transcript number in each species.}
}
\value{
A \code{arma::vec} represents the total estimated counts of each species.
}
\description{
Estimated counts of input species.
}
\details{
The indices of \code{est} should be consistent with the \code{spenum}. For example, the \code{est} is \code{1.5, 2, 3} and \code{spenum} is \code{2, 1}, so \code{2.5, 3} will be returned.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
