% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{EM}
\alias{EM}
\title{EM model}
\usage{
EM(pseudo, spenum, maxiter = 10000, n = 2)
}
\arguments{
\item{pseudo}{A \code{list} of equivalence class (ec), count of ec, and effective length of transcripts. It can be the outputs of kallisto.}

\item{spenum}{A \code{arma::uvec} indicated number of transcripts.}

\item{maxiter}{The maximum iteration number with the default value of 10000.}

\item{n}{The number of CPUs or processors.}
}
\value{
A \code{numeric vector} indicates probabilities of selecting a read from the different transcripts.
}
\description{
Expectation maximization (EM) model for RNA-seq quantification.
}
\details{
EM model for RNA-seq quantification.
}
\examples{
##    f1 f2 f3
## ec1 1 1 1
## ec2 0 1 1
## ec3 1 0 1
## ec4 1 0 0
## ec5 1 1 0
plist <- list(ec = c('0,1,2', '1,2', '0,2', '0', '0,1'), count = rep(1, 5), efflen = rep(1, 3))
EM(plist, spenum = 3, n = 2)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
