% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMSingle}
\alias{EMSingle}
\title{Single EM iteration}
\usage{
EMSingle(prob, efflen, ec, count)
}
\arguments{
\item{prob}{A \code{arma::vec} of probabilities of selecting a read from the different transcripts.}

\item{efflen}{A \code{std::vector<arma::vec>} indicated the effective length of transcripts.}

\item{ec}{A \code{std::vector<arma::vec>} indicated equivalence classes with the same length of \code{efflen}.}

\item{count}{A \code{arma::uvec} indicated the mapped count number of equivalence class with the same length of \code{efflen}.}
}
\value{
A updated \code{arma::vec} estimated counts of different transcripts.
}
\description{
Parallel a single EM iteration.
}
\details{
One iteration of EM model.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
