% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogSumExp}
\alias{LogSumExp}
\alias{LogSumExp1}
\alias{LogSumExpRatio}
\alias{LogSumExpRatio1}
\title{LogSumExp}
\usage{
LogSumExp(x, weight)

LogSumExp1(x)

LogSumExpRatio(x, weight)

LogSumExpRatio1(x)
}
\arguments{
\item{x}{A \code{arma::vec}.}

\item{weight}{A \code{arma::vec} indicating the weight.}
}
\description{
Calculate the log-sum-exp calculator
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
